update tests for moved classes
